requirements.txt
requests==2.9.1
gensim==3.3.0
PyStemmer==1.3.0
